Change nbytes to 64 bit

+++ erts/emulator/drivers/common/erl_efile.h
@@ -163,4 +163,4 @@ int efile_may_openfile(Efile_error* errInfo, char *name);
 int efile_fadvise(Efile_error* errInfo, int fd, Sint64 offset, Sint64 length,
 		  int advise);
 int efile_sendfile(Efile_error* errInfo, int in_fd, int out_fd,
-		      off_t *offset, size_t *nbytes);
+		      off_t *offset, Uint64 *nbytes);
